+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Migration to RJW 5.5.0, big thanks to @prototype99 (#146)
 - Removed some Genitalia and Genes that substituted Race-Support. Golem, Pig, Ghost, Marine and Orc Genitalia are gone (for now), but an enthusiast can surely bring it back.
 - Evergrowth is not exclusive with normal genitalia-sizes anymore. 
+- "Very much Cum" Gene -> "Very much Fluid" Gene. Works for all primary genitalia of both genders now.  
 
 **Fixes**: